requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.4=pyhd8ed1ab_0
arviz=0.11.0=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
bzip2=1.0.8=hfdf4475_7
ca-certificates=2024.8.30=h8857fd0_0
certifi=2024.8.30=pypi_0
cftime=1.6.4=pypi_0
comm=0.2.2=pyhd8ed1ab_0
cycler=0.12.1=pypi_0
debugpy=1.8.6=py310h53e7c6a_0
decorator=5.1.1=pyhd8ed1ab_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.1.0=pyhd8ed1ab_0
fonttools=4.54.1=pypi_0
importlib-metadata=8.5.0=pyha770c72_0
ipykernel=6.29.5=pyh57ce528_0
ipython=8.27.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jupyter_client=8.6.3=pyhd8ed1ab_0
jupyter_core=5.7.2=pyh31011fe_1
kiwisolver=1.4.7=pypi_0
krb5=1.21.3=h37d8d59_0
libcxx=19.1.0=hf95d169_0
libedit=3.1.20191231=h0678c8f_2
libffi=3.4.2=h0d85af4_5
libsodium=1.0.20=hfdf4475_0
libsqlite=3.46.1=h4b8f8c9_0
libzlib=1.3.1=h87427d6_1
llvmlite=0.43.0=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
ncurses=6.5=hf036a51_1
nest-asyncio=1.6.0=pyhd8ed1ab_0
netcdf4=1.7.1.post2=pypi_0
numba=0.60.0=pypi_0
numpy=1.22.4=pypi_0
openssl=3.3.2=hd23fc13_0
packaging=24.1=pyhd8ed1ab_0
pandas=2.2.3=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.4.0=pypi_0
pip=24.2=pyh8b19718_1
platformdirs=4.3.6=pyhd8ed1ab_0
prompt-toolkit=3.0.48=pyha770c72_0
psutil=6.0.0=py310h837254d_1
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.3=pyhd8ed1ab_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.1.4=pypi_0
python=3.10.15=h6bafeb6_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python_abi=3.10=5_cp310
pytz=2024.2=pypi_0
pyzmq=26.2.0=py310hfe2c05e_2
readline=8.2=h9e318b2_1
scipy=1.7.3=pypi_0
seaborn=0.13.2=pypi_0
setuptools=75.1.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
stack_data=0.6.2=pyhd8ed1ab_0
tk=8.6.13=h1abcd95_1
tornado=6.4.1=py310h837254d_1
tqdm=4.66.5=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
typing-extensions=3.10.0.2=pypi_0
tzdata=2024.2=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.44.0=pyhd8ed1ab_0
xarray=2023.12.0=pypi_0
xz=5.2.6=h775f41a_0
zeromq=4.3.5=hb33e954_5
zipp=3.20.2=pyhd8ed1ab_0
